Move release from run in common
